feat: Add movement speed stat

+++ Assets/Scripts/Utilities/Stats/StatData.cs
@@ -1,5 +1,5 @@
 namespace Stats
 {
     public record StatData(float BaseValue, float AdditiveValue = 0, float MultiplicativeValue = 1);
-    public record ResourceData(StatData Max, int Value);
+    public record ResourceData(StatData Max, float Value);
 }
